+++ test/clojure-mode-refactor-rename-ns-alias-test.el
@@ -15,6 +15,11 @@
 ;; You should have received a copy of the GNU General Public License
 ;; along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+
+;;; Commentary:
+
+;; Tests for clojure-rename-ns-alias
+
 ;;; Code:
 
 (require 'clojure-mode)